Mavenize and reorganize as per discussion in https://github.com/ipfs/java-ipfs-api/issues/1
