+++ .gitignore
@@ -1,5 +1,6 @@
 *.class
 
+target/**
 build/**
 dist/**
 .idea/**
